requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
alembic==1.13.3
pytest==8.3.3
httpx==0.27.2
python-jose==3.3.0
python-multipart==0.0.16
jinja2==3.1.4
aiofiles==24.1.0
pydantic[email]
passlib[bcrypt]